requirements.txt
anthropic==0.30.1
openai~=1.35.7
PyGithub~=2.3.0
aider-chat